requirements.txt
Flask==2.3.3
requests==2.31.0
PyYAML==6.0
gitpython==3.1.31
pytest==7.4.0
fastapi==0.100.0
uvicorn[standard]==0.23.0
python-dotenv==1.0.1
